Change datatype of vaiable in function

+++ src/Customer.java
@@ -42,8 +42,8 @@ class Customer {
 		return result;
 	}
 
-	private int amountFor(Rental each) {
-		int thisAmount = 0;
+	private double amountFor(Rental each) {
+		double thisAmount = 0;
 		switch (each.getMovie().getPriceCode()) {
 		case Movie.REGULAR:
 			thisAmount += 2;
